pyproject.toml
[project]
name = "the-gatherer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.7",
    "django-stubs>=5.1.3",
    "djangorestframework>=3.15.2",
    "djangorestframework-stubs>=3.15.3",
    "python-dotenv>=1.1.0",
]
